+++ includes/math.h
@@ -6,7 +6,7 @@
 /*   By: ebaudet <ebaudet@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2014/04/24 19:51:44 by ebaudet           #+#    #+#             */
-/*   Updated: 2019/03/20 17:47:11 by ebaudet          ###   ########.fr       */
+/*   Updated: 2019/04/29 18:16:30 by ebaudet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,6 +16,5 @@
 long long int	ft_pow(int nb, unsigned int pow);
 void			ft_puthex(long long int nb, int width);
 char			*ft_gethex(long long int nb, int width);
-
 
 #endif
