+++ libft/ft_filewithanameverylongtotestwhatisthecomportementofthefunctionnmewiththearchive.c
@@ -1,23 +1,22 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_strcmp.c                                        :+:      :+:    :+:   */
+/*   ft_filewithanameverylongtotestwhatisthecomportementofthefunctionnmewiththearchive.c:+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: ebaudet <ebaudet@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2013/11/19 16:49:07 by ebaudet           #+#    #+#             */
-/*   Updated: 2019/04/29 18:18:33 by ebaudet          ###   ########.fr       */
+/*   Created: 2019/03/19 11:06:05 by ebaudet           #+#    #+#             */
+/*   Updated: 2019/04/29 18:14:04 by ebaudet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
-
 #include "libft.h"
 
-int		ft_strcmp(const char *s1, const char *s2)
-{
-	int		i;
+/*
+** The bzero() function writes <n> zeroed bytes to the string <s>.
+** If <n> is zero, bzero() does nothing.
+*/
 
-	i = 0;
-	while (s1[i] == s2[i] && s1[i] != '\0' && s2[i] != '\0')
-		i++;
-	return ((unsigned char)s1[i] - (unsigned char)s2[i]);
+void	ft_filewithanameverylongtotestwhatisthecomportementofthefunctionnmewiththearchive(void *s, size_t n)
+{
+	ft_memset(s, 0, n);
 }
