+++ libft/ft_ldtoa.c
@@ -6,7 +6,7 @@
 /*   By: ebaudet <ebaudet@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2013/11/19 16:39:09 by ebaudet           #+#    #+#             */
-/*   Updated: 2019/02/12 19:00:34 by ebaudet          ###   ########.fr       */
+/*   Updated: 2019/04/29 18:19:33 by ebaudet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
