+++ libft/includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: ebaudet <ebaudet@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2013/11/21 11:44:10 by ebaudet           #+#    #+#             */
-/*   Updated: 2019/03/19 11:06:41 by ebaudet          ###   ########.fr       */
+/*   Updated: 2019/04/29 18:10:55 by ebaudet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -99,7 +99,8 @@ char	*ft_ldtoa(long double d, unsigned int precision);
 int		ft_max(int a, int b);
 int		ft_min(int a, int b);
 
-void	ft_filewithanameverylongtotestwhatisthecomportementofthefunctionnmewiththearchive(void *s, size_t n);
+void	ft_filewithanameverylongtotestwhatisthecomportementofthefunctionnmewith\
+thearchive(void *s, size_t n);
 
 /*
 ** LIBFT_H
